move util_ functions to datajudge.utils
