+++ src/datajudge/requirements.py
@@ -279,10 +279,10 @@ class WithinRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
 
@@ -328,10 +328,10 @@ class WithinRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
 
@@ -388,10 +388,10 @@ class WithinRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
 
@@ -1489,10 +1489,10 @@ class BetweenRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
 
@@ -1541,10 +1541,10 @@ class BetweenRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
 
@@ -1603,10 +1603,10 @@ class BetweenRequirement(Requirement):
         for ``WithinRequirement``.
         By default, the null filtering does not trigger if multiple columns are fetched at once.
         It can be configured in more detail by supplying a custom ``filter_func`` function.
-        Some exemplary implementations are available in this module as ``util_filternull_default_deprecated``,
-        ``util_filternull_never``, ``util_filternull_element_or_tuple_all``, ``util_filternull_element_or_tuple_any``.
+        Some exemplary implementations are available in this module as ``datajudge.utils.util_filternull_default_deprecated``,
+        ``datajudge.utils.util_filternull_never``, ``datajudge.utils.util_filternull_element_or_tuple_all``, ``datajudge.utils.util_filternull_element_or_tuple_any``.
         For new deployments, using one of the above filters or a custom one is recommended.
-        Passing None as the argument is equivalent to ``util_filternull_default_deprecated``, but triggers a warning.
+        Passing None as the argument is equivalent to ``datajudge.utils.util_filternull_default_deprecated``, but triggers a warning.
         The deprecated default may change in future versions.
         To silence the warning, set ``filter_func`` explicitly.
